Added needed screens for bluetooth connection
