+++ app/human-detector-app/config/BLEConfig.ts
@@ -1,0 +1,4 @@
+export const BLE_SERVICE_UUID = '4539d44c-75bb-4fbd-9d95-6cdf49d4ef2b';
+export const WIFI_UUID = '7a1673f9-55cb-4f40-b624-561ad8afb8e2';
+export const SERIAL_UUID = '8b83fee2-373c-46a5-a782-1db9118431d9';
+export const CONNECTION_UUID = '136670fb-f95b-4ee8-bc3b-81eadb234268';
