+++ app/human-detector-app/package.json
@@ -46,7 +46,7 @@
     "expo-status-bar": "~1.4.0",
     "expo-system-ui": "~1.3.0",
     "expo-updates": "~0.14.6",
-    "expo-web-browser": "~12.0.0",
+    "expo-web-browser": "~11.0.0",
     "node-fetch": "^3.2.10",
     "react": "18.0.0",
     "react-dom": "18.0.0",
