+++ app/human-detector-app/screens/CameraRegistration/EnterCameraRegInfoScreen.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+import { View } from 'react-native';
+
+/**
+ * The EnterCameraRegInfoScreen will allow the user
+ * to enter the information that they will need for the mobile app camera registration.
+ * This information includes:
+ *  WIFI Password,
+ *  Camera Name
+ */
+export default function EnterCameraRegInfoScreen(): React.ReactElement {
+    return (
+        <View/>
+    )
+}
